crash_service/requirements.txt
fastapi
python-dotenv
uvicorn
gunicorn
pymongo